PredicateDefinitionCompiler
