+++ crates/core/src/pattern_compiler/mod.rs
@@ -29,6 +29,7 @@ mod node_compiler;
 pub(crate) mod not_compiler;
 pub(crate) mod or_compiler;
 pub(crate) mod pattern_definition_compiler;
+pub(crate) mod predicate_definition_compiler;
 pub(crate) mod step_compiler;
 
 pub(crate) use compiler::{parse_one, CompilationContext};
